+++ libs/comms/src/CClientTCPSocket.cpp
@@ -32,6 +32,7 @@
 #include <netdb.h>
 #include <netinet/in.h>
 #include <netinet/tcp.h>
+#include <sys/epoll.h>	// epoll_create1()
 #include <sys/ioctl.h>
 #include <sys/socket.h>
 #include <sys/types.h>
@@ -65,7 +66,13 @@ CClientTCPSocket::CClientTCPSocket()
 	m_hSock = INVALID_SOCKET;
 #else
 	// Linux, Apple
-	m_hSock = -1;
+	m_epoll4read_fd = epoll_create1(0);
+	m_epoll4write_fd = epoll_create1(0);
+
+	ASSERTMSG_(
+		m_epoll4read_fd != -1 && m_epoll4write_fd != -1,
+		"[CClientTCPSocket] Failed to create epoll file descriptor!");
+
 #endif
 	MRPT_END
 }
@@ -84,7 +91,9 @@ CClientTCPSocket::~CClientTCPSocket()
 #ifdef _WIN32
 	WSACleanup();
 #else
-// Nothing else to do.
+	// Clean up epoll:
+	if (m_epoll4read_fd != -1) ::close(m_epoll4read_fd);
+	if (m_epoll4write_fd != -1) ::close(m_epoll4write_fd);
 #endif
 }
 
@@ -104,6 +113,9 @@ void CClientTCPSocket::close()
 	// Delete socket:
 	if (m_hSock != -1)
 	{
+		epoll_ctl(m_epoll4write_fd, EPOLL_CTL_DEL, m_hSock, nullptr);
+		epoll_ctl(m_epoll4read_fd, EPOLL_CTL_DEL, m_hSock, nullptr);
+
 		shutdown(m_hSock, SHUT_RDWR);
 		::close(m_hSock);
 		m_hSock = -1;
@@ -140,6 +152,22 @@ void CClientTCPSocket::sendString(const std::string& str)
 {
 	Write(str.c_str(), str.size());
 }
+
+#ifndef _WIN32
+void CClientTCPSocket::internal_attach_epoll_to_hsock()
+{
+	struct epoll_event event;
+	event.data.fd = m_hSock;
+
+	event.events = EPOLLOUT;  // EPOLLERR is always waited for.
+	if (0 != epoll_ctl(m_epoll4write_fd, EPOLL_CTL_ADD, m_hSock, &event))
+		THROW_EXCEPTION("epoll_ctl() for write events returned error.");
+
+	event.events = EPOLLIN;
+	if (0 != epoll_ctl(m_epoll4read_fd, EPOLL_CTL_ADD, m_hSock, &event))
+		THROW_EXCEPTION("epoll_ctl() for read events returned error.");
+}
+#endif
 
 /*---------------------------------------------------------------
 						connect
@@ -204,29 +232,28 @@ void CClientTCPSocket::connect(
 			"Error connecting to %s:%hu. Error: %s [%d]",
 			remotePartAddress.c_str(), remotePartTCPPort, strerror(er), er));
 
+	// socket is created:
+	internal_attach_epoll_to_hsock();
+
 	// Wait for connect:
-	timeval timer = {0, 0};
-	fd_set ss_write, ss_errors;
-	FD_ZERO(&ss_write);
-	FD_ZERO(&ss_errors);
-	FD_SET(m_hSock, &ss_write);
-	FD_SET(m_hSock, &ss_errors);
 
-	timer.tv_sec = timeout_ms / 1000;
-	timer.tv_usec = 1000 * (timeout_ms % 1000);
+	std::array<struct epoll_event, 1> events;
 
-	int sel_ret = select(
-		m_hSock + 1,
-		nullptr,  // For read
-		&ss_write,	// For write or *connect done*
-		&ss_errors,	 // For errors
-		timeout_ms == 0 ? nullptr : &timer);
+	const int epoll_timeout_ms = timeout_ms == 0 ? -1 : timeout_ms;
 
-	if (sel_ret == 0)
+	int event_count;
+	do
+	{
+		event_count = epoll_wait(
+			m_epoll4write_fd, events.data(), events.size(), epoll_timeout_ms);
+	} while (event_count < 0 && errno == EINTR);
+
+	if (event_count == 0)
 		THROW_EXCEPTION(format(
 			"Timeout connecting to '%s:%hu':\n%s", remotePartAddress.c_str(),
 			remotePartTCPPort, getLastErrorStr().c_str()));
-	if (sel_ret == -1)
+
+	if (event_count == -1)
 		THROW_EXCEPTION(format(
 			"Error connecting to '%s:%hu':\n%s", remotePartAddress.c_str(),
 			remotePartTCPPort, getLastErrorStr().c_str()));
@@ -290,13 +317,7 @@ size_t CClientTCPSocket::readAsync(
 	int readNow;
 	bool timeoutExpired = false;
 
-	struct timeval timeoutSelect = {0, 0};
-	struct timeval* ptrTimeout;
-	fd_set sockArr;
-
-	// Init fd_set structure & add our socket to it:
-	FD_ZERO(&sockArr);
-	FD_SET(m_hSock, &sockArr);
+	std::array<struct epoll_event, 1> events;
 
 	// Loop until timeout expires or the socket is closed.
 	while (alreadyRead < Count && !timeoutExpired)
@@ -304,27 +325,23 @@ size_t CClientTCPSocket::readAsync(
 		// Use the "first" or "between" timeouts:
 		int curTimeout = alreadyRead == 0 ? timeoutStart_ms : timeoutBetween_ms;
 
-		if (curTimeout < 0) ptrTimeout = nullptr;
-		else
-		{
-			timeoutSelect.tv_sec = curTimeout / 1000;
-			timeoutSelect.tv_usec = 1000 * (curTimeout % 1000);
-			ptrTimeout = &timeoutSelect;
-		}
+		const int epoll_timeout_ms =
+			(curTimeout < 0) ? -1 /*No timeout*/ : curTimeout;
 
 		// Wait for received data
-		int selRet = ::select(
-			m_hSock + 1,  // __nfds
-			&sockArr,  // Wait for read
-			nullptr,  // Wait for write
-			nullptr,  // Wait for except.
-			ptrTimeout);  // Timeout
+		int event_count;
+		do
+		{
+			event_count = epoll_wait(
+				m_epoll4read_fd, events.data(), events.size(),
+				epoll_timeout_ms);
+		} while (event_count < 0 && errno == EINTR);
 
-		if (selRet == INVALID_SOCKET)
+		if (event_count < 0)
 			THROW_EXCEPTION_FMT(
 				"Error reading from socket: %s", getLastErrorStr().c_str());
 
-		if (selRet == 0)
+		if (event_count == 0)
 		{
 			// Timeout:
 			timeoutExpired = true;
@@ -379,39 +396,29 @@ size_t CClientTCPSocket::writeAsync(
 	int writtenNow;
 	bool timeoutExpired = false;
 
-	struct timeval timeoutSelect = {0, 0};
-	struct timeval* ptrTimeout;
-	fd_set sockArr;
-
-	// Init fd_set structure & add our socket to it:
-	FD_ZERO(&sockArr);
-	FD_SET(m_hSock, &sockArr);
+	std::array<struct epoll_event, 1> events;
 
 	// The timeout:
-	if (timeout_ms < 0) { ptrTimeout = nullptr; }
-	else
-	{
-		timeoutSelect.tv_sec = timeout_ms / 1000;
-		timeoutSelect.tv_usec = 1000 * (timeout_ms % 1000);
-		ptrTimeout = &timeoutSelect;
-	}
+	const int epoll_timeout_ms = timeout_ms < 0 ? -1 : timeout_ms;
 
 	// Loop until timeout expires or the socket is closed.
 	while (alreadyWritten < Count && !timeoutExpired)
 	{
 		// Wait for received data
-		int selRet = ::select(
-			m_hSock + 1,  // __nfds
-			nullptr,  // Wait for read
-			&sockArr,  // Wait for write
-			nullptr,  // Wait for except.
-			ptrTimeout);  // Timeout
+		int event_count;
+		do
+		{
+			event_count = epoll_wait(
+				m_epoll4write_fd, events.data(), events.size(),
+				epoll_timeout_ms);
 
-		if (selRet == INVALID_SOCKET)
+		} while (event_count < 0 && errno == EINTR);
+
+		if (event_count < 0)
 			THROW_EXCEPTION_FMT(
 				"Error writing to socket: %s", getLastErrorStr().c_str());
 
-		if (selRet == 0)
+		if (event_count == 0)
 		{
 			// Timeout:
 			timeoutExpired = true;